pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-kitab-diff"
version = "0.1.0"
description = "A diff tool for historical texts, based on WikEdDiff."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Peter Verkinderen" }]
dependencies = ["python-wikeddiff"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]